Add option to search to restrict to branch/child terms
